+++ src/test/kotlin/com/zepben/auth/client/ZepbenTokenFetcherTest.kt
@@ -94,6 +94,31 @@ internal class ZepbenTokenFetcherTest {
     }
 
     @Test
+    fun testCreateTokenFetcherNoTokenPath() {
+        doReturn(StatusCode.OK.code).`when`(response).statusCode()
+        doReturn(
+            "{\"authType\": \"OAUTH\", \"audience\": \"test_audience\", \"issuerDomain\": \"test_issuer\"}"
+        ).`when`(response).body()
+
+        val tokenFetcher = createTokenFetcher("https://testaddress", confClient = client, authClient = client)
+        verify(client).send(any(), any<HttpResponse.BodyHandler<String>>())
+        assertThat(tokenFetcher?.audience, equalTo("test_audience"))
+        assertThat(tokenFetcher?.issuerDomain, equalTo("test_issuer"))
+    }
+
+    @Test
+    fun testCreateTokenFetcherLowercaseAuthType() {
+        doReturn(StatusCode.OK.code).`when`(response).statusCode()
+        doReturn(
+            "{\"authType\": \"oauth\", \"audience\": \"test_audience\", \"issuerDomain\": \"test_issuer\", \"tokenPath\": \"/oauth/token\"}"
+        ).`when`(response).body()
+
+        val tokenFetcher = createTokenFetcher("https://testaddress", confClient = client, authClient = client)
+        verify(client).send(any(), any<HttpResponse.BodyHandler<String>>())
+        assertThat(tokenFetcher?.authMethod, equalTo(AuthMethod.OAUTH))
+    }
+
+    @Test
     fun testCreateTokenFetcherNoAuth() {
         doReturn(StatusCode.OK.code).`when`(response).statusCode()
         doReturn(